Validations for login, register, ForgotPasswordModel Update DB

+++ src/main/java/com/sda/javaremote18/spring_boot/controllers/UsersRepository.java
@@ -4,7 +4,7 @@ import com.sda.javaremote18.spring_boot.models.UserModel;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsersRepository extends CrudRepository<UserModel, Long> {
+public interface UsersRepository extends CrudRepository<UserModel, Integer> {
 
     @Query("SELECT user from UserModel user where user.email=?1")
     public UserModel findByEmail(String email);
